refactor: updated imports of userActions
