+++ src/asyncActions/categories.ts
@@ -1,7 +1,7 @@
 import { Dispatch } from "../../node_modules/redux/index";
-import {getCategoriesAction} from "../store/categoryReducer"
 import {CategoryAction} from "./../types/types"
 import categories from './../data/categories.json';
+import { getCategoriesAction } from "../store/userActions";
 
 const emulateAPI = () =>
   new Promise((resolve) => {
